Refactor in progress 2
