+++ typescript/src/lib/extractors/jira/components/query-builder.ts
