+++ typescript/src/lib/extractors/jira/extractor.ts
@@ -1,6 +1,10 @@
-import { getAllWorkItemsFromJiraApi } from './api-adapter/main';
 import { IJiraSettings, JiraSettings } from './settings'
-import { IWorkItem } from '../../core/work-item';
+import { IIssueList, IIssue } from './models';
+import { buildJiraQueryUrl } from './components/query-builder';
+import { getAttributes } from './components/attribute-parser';
+import { IWorkItem, WorkItem } from '../../core/work-item';
+import { getJsonFromUrl, getHeaders } from '../../core/http';
+import { getStagingDates } from './components/staging-parser';
 
 class JiraExtractor {
   private settings: IJiraSettings;
@@ -27,7 +31,7 @@ class JiraExtractor {
 
   getWorkItems(): Promise<any> {
     return new Promise((accept, reject) => {
-      getAllWorkItemsFromJiraApi(this.settings, this.statusHook)
+      this.getAllWorkItemsFromJiraApi(this.settings, this.statusHook)
       .then(items => {
         this.workItems = items;
         accept(items);
@@ -37,32 +41,94 @@ class JiraExtractor {
     });
   };
 
-  getWorkItemBatch(): Promise<any> {
-    return null;
+  testConnectionToJira = async function(settings: IJiraSettings): Promise<number> {
+    const metadata = await this.getWorkItemMetadata(settings);
+    return metadata.totalJiras ? metadata.totalJiras : false;
   };
 
-  testConnectionToJira(): Promise<any> {
-    return null;
-  }
-
-  getWorkItemMetadata(): Promise<any> {
-    return null;
-  }
-
+  getWorkItemMetadata = async function(settings: IJiraSettings): Promise<any> {
+    const metadata = await getJsonFromUrl(
+        buildJiraQueryUrl(
+          settings.ApiUrl, 
+          settings.Criteria.Projects, 
+          settings.Criteria.IssueTypes, 
+          settings.Criteria.Filters), 
+        getHeaders(settings.Connection.Username, settings.Connection.Password)
+    );
+    return metadata;
+  };
 
   toCSV(workItems = this.workItems, stages = this.settings.Stages, attributes = this.settings.Attributes): string {
     const header = `ID,Link,Name,${stages.join(',')},Type,${Object.keys(attributes).join(',')}`;
     const body = workItems.map(item => item.toCSV()).reduce((res, cur) => `${res + cur}\n`, '');
     const csv = `${header}\n${body}`;
     return csv;
-  }
+  };
 
   toSerializedArray(workItems = this.workItems, stages = this.settings.Stages, attributes = this.settings.Attributes): string {
     const header = `["ID","Link","Name",${stages.map(stage => `"${stage}"`).join(',')},"Type",${Object.keys(attributes).map(attribute => `"${attribute}"`).join(',')}]`;
     const body = workItems.map(item => item.toSerializedArray()).reduce((res, cur) => `${res},\n${cur}`, '');
     const serializedData: string = `[${header}${body}]`;
     return serializedData;
-  }
+  };
+
+  getIssues = async function(query: string, username: string, password: string): Promise<IIssue[]> {
+    const headers: Headers = getHeaders(username, password);
+    const result: IIssueList = await getJsonFromUrl(query, headers);
+    if (result.issues) {
+        const issues: IIssue[] = result.issues;
+        return issues;
+    } else {
+        throw new Error('Could not retrieve issues from object');
+    }
+  };
+
+  createWorkItem(issue: IIssue, settings: IJiraSettings): IWorkItem {
+    const key: string = issue.key;
+    const name: string = issue.fields['summary'];
+    // refactor this one
+    const stagingDates = getStagingDates(issue, settings.Stages, settings.StageMap, settings.CreateInFirstStage, settings.ResolvedInLastStage);
+    const type = issue.fields.issuetype.name ? issue.fields.issuetype.name : '';
+
+    const requestedAttributeSystemNames: string[] = Object.keys(settings.Attributes).map(key => settings.Attributes[key]);
+    const attributes = getAttributes(issue.fields, requestedAttributeSystemNames);
+
+    const workItem = new WorkItem(key, stagingDates, name, type, attributes);
+    return workItem;
+  };
+
+  getWorkItemsBatch = async function(start: number, batchSize: number, settings: IJiraSettings): Promise<IWorkItem[]> {
+    const url = buildJiraQueryUrl(
+      settings.ApiUrl, 
+      settings.Criteria.Projects, 
+      settings.Criteria.IssueTypes, 
+      settings.Criteria.Filters, 
+      start, 
+      batchSize
+    );
+    const issues = await this.getIssues(url, settings.Connection.Username, settings.Connection.Password);
+    const workItems = issues.map(issue => this.createWorkItem(issue, settings));
+    return workItems;
+  };
+
+  getAllWorkItemsFromJiraApi = async function(settings: IJiraSettings, hook: any = () => {}, resultsPerBatch = 25): Promise<IWorkItem[]> {
+    const metadata = await this.getWorkItemMetadata(settings);
+    const totalJiras: number = metadata.total; 
+    const batchSize: number = resultsPerBatch;
+    const totalBatches: number = Math.ceil(totalJiras / batchSize); 
+
+    const allWorkItems: IWorkItem[] = [];
+    hook(0);
+    for  (let i = 0; i < totalBatches; i++) {
+      const workItemBatch = await this.getWorkItemsBatch(i * batchSize, batchSize, settings);
+      allWorkItems.push(...workItemBatch);
+      hook(Math.max(batchSize / totalJiras)*100);
+    }
+    hook(100);
+
+    return allWorkItems;
+  };
+
 };
 
 export {
